rename and move package to smile.feature
